This reverts commit 689951708fccd0073873e645bf683bc34b75a772 This reverts commit da70f8047d3f9c8af9d3eee3b14c71122c5220a0.
Reverting v3 and v2 so that we can take a simpler approach that touches far fewer files.
