+++ src/hotspot/cpu/x86/macroAssembler_x86.cpp
@@ -46,7 +46,7 @@
 #include "runtime/safepointMechanism.hpp"
 #include "runtime/sharedRuntime.hpp"
 #include "runtime/stubRoutines.hpp"
-#include "runtime/thread.inline.hpp"
+#include "runtime/thread.hpp"
 #include "utilities/macros.hpp"
 #include "crc32c.h"
