+++ src/hotspot/share/aot/aotLoader.cpp
@@ -32,7 +32,6 @@
 #include "oops/method.hpp"
 #include "runtime/arguments.hpp"
 #include "runtime/handles.inline.hpp"
-#include "runtime/thread.inline.hpp"
 #include "runtime/os.inline.hpp"
 #include "runtime/timerTrace.hpp"
