+++ src/hotspot/share/gc/shared/concurrentGCBreakpoints.cpp
@@ -29,7 +29,7 @@
 #include "memory/universe.hpp"
 #include "runtime/interfaceSupport.inline.hpp"
 #include "runtime/mutexLocker.hpp"
-#include "runtime/thread.inline.hpp"
+#include "runtime/thread.hpp"
 #include "utilities/debug.hpp"
 
 // States:
