+++ src/hotspot/share/jfr/support/jfrEpochSynchronization.cpp
@@ -1,5 +1,5 @@
 /*
-* Copyright (c) 2019, 2020, Oracle and/or its affiliates. All rights reserved.
+* Copyright (c) 2019, Oracle and/or its affiliates. All rights reserved.
 * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
 *
 * This code is free software; you can redistribute it and/or modify it
